Moved test_reprocessFileList_processesInOrder to a periodic/serial test suite It looks like this was a race condition due to tests running in parallel
